apps/worker/requirements.txt
APScheduler==3.10.4
asyncpg==0.30.0
httpx==0.28.1
pytest==8.3.5
SQLAlchemy==2.0.40
python-dotenv==1.0.1
redis==5.2.1
